requirements.txt
streamlit==1.42.2
opencv-python==4.10.0.84
deepface==0.0.90
pillow==10.2.0
numpy==1.26.4
